fixed a bunch of compiler warnings

+++ microcontroller/src-atmel/borg/borgware-2d/display_loop.c
@@ -53,7 +53,7 @@ void display_loop(){
 
 #ifdef RANDOM_SUPPORT
 			{ char a[28];
-				sprintf(a,"</# counter == %lu  ", percnt_get());
+				sprintf(a,"</# counter == %lu  ", (unsigned long) percnt_get());
 				scrolltext(a);
 			}
 #endif
